Vafhilhi
the following copyright notice and this permission notice shall be included
tags in the documentation and/or other materials provided with the distribution
language governing permissions and limitations there under. This software shall
link to the original copyright notice and this permission notice shall be included
highlighted in all copies or substantial portions of the software.